Added Auto migration Options for Core Data.

+++ Category.Foundation/NSString+Additions.m
@@ -4,7 +4,7 @@
 - (NSString *)URLEncodedString {
     NSMutableString * output = [NSMutableString string];
     const unsigned char * source = (const unsigned char *)[self UTF8String];
-    int sourceLen = strlen((const char *)source);
+    size_t sourceLen = strlen((const char *)source);
     for (int i = 0; i < sourceLen; ++i) {
         const unsigned char thisChar = source[i];
         if (thisChar == ' '){
